requirements.txt
pandas
requests
python-dotenv
nltk
spacy
tqdm
streamlit
faiss-cpu
langchain
openai
